requirements.txt
nextcord
asyncio
pytz
psutil
discordmongopy==0.3.0
motor
python-dotenv==0.15.0
datetime
requests
pymongo
dnspython==2.1.0
dataclasses
python-dateutil==2.8.1
aiohttp
Pillow
asyncio
urllib3==1.26.3
bs4
beautifulsoup4==4.9.3
pyyaml
schedule